+++ Assets/Scripts/Clusterization/Algorithms/Simple3PlaneClusterization.cs
@@ -10,17 +10,6 @@ namespace Elektronik.Clusterization.Algorithms
         public Simple3PlaneClusterization(Vector3 offset)
         {
             _offset = offset;
-            _colors = new[]
-            {
-                Color.blue,
-                Color.red,
-                Color.green,
-                Color.yellow,
-                Color.magenta,
-                new Color(0.5f, 0.5f, 1f),
-                new Color(1f, 0.5f, 0),
-                new Color(0.5f, 0, 1)
-            };
         }
 
         public List<List<SlamPoint>> Compute(IList<SlamPoint> items)
@@ -51,21 +40,10 @@ namespace Elektronik.Clusterization.Algorithms
                                          && i.Position.z < _offset.z).ToList();
 
             var res = new List<List<SlamPoint>> {l1, l2, l3, l4, l5, l6, l7, l8};
-            for (int i = 0; i < res.Count; i++)
-            {
-                for (int j = 0; j < res[i].Count; j++)
-                {
-                    var slamPoint = res[i][j];
-                    slamPoint.Color = _colors[i];
-                    res[i][j] = slamPoint;
-                }
-            }
 
             return res;
         }
 
         private readonly Vector3 _offset;
-
-        private readonly Color[] _colors;
     }
 }