+++ Assets/Scripts/EditorTests/DataRecorderSubscription.cs
@@ -156,11 +156,5 @@ namespace Elektronik.EditorTests
             mockedRecorder.Verify(r => r.OnRemoved<SlamObservation>(It.IsAny<string>(),  It.IsAny<List<int>>()),
                                    Times.Once());
         }
-
-        [Test]
-        public void FailedTest()
-        {
-            Assert.Fail();
-        }
     }
 }