+++ plugins/Protobuf/Online/GrpcServices/ImageManager.cs
@@ -1,15 +1,18 @@
 ﻿using System;
 using System.Diagnostics;
+using System.Security;
 using System.Threading.Tasks;
 using Elektronik.Protobuf.Data;
 using Elektronik.Protobuf.Online.Presenters;
 using Grpc.Core;
-using Debug = UnityEngine.Debug;
+using Grpc.Core.Logging;
 
 namespace Elektronik.Protobuf.Online.GrpcServices
 {
     public class ImageManager : ImageManagerPb.ImageManagerPbBase
     {
+        public ILogger Logger = new UnityLogger();
+
         private readonly RawImagePresenter _presenter;
 
         public ImageManager(RawImagePresenter presenter)
@@ -33,9 +36,10 @@ namespace Elektronik.Protobuf.Online.GrpcServices
                 err.ErrType = ErrorStatusPb.Types.ErrorStatusEnum.Failed;
                 err.Message = e.Message;
             }
+
             timer.Stop();
-            Debug.Log($"[{GetType().Name}.Handle] Elapsed time: {timer.ElapsedMilliseconds} ms. " +
-                      $"ErrorStatus: {err}");
+            Logger.Info($"[{GetType().Name}.Handle] Elapsed time: {timer.ElapsedMilliseconds} ms. ErrorStatus: {err}");
+
             return Task.FromResult(err);
         }
     }