+++ ReleaseNotes.md
@@ -1,6 +1,15 @@
 # Новое в этой версии:
-- Запись последовательностей
+- Обнаружение плоскостей в облаке точек
+- Возможность выключить/включить логирование в файл
+- Возможность передавать данные на другие запущенные экземпляры приложения
+- Возможность менять прозрачность линий
+- Камера может следовать за отслеживаемым объектом
+- Оптимизировано отображение наблюдений
 
 # New in this version:
-- Recording of incoming data
-
+- Planes detection tool
+- User can disable/enable logging in file
+- Plugin that allows Elektronik to transmit data from one instance to another
+- Semitransparent lines
+- Camera can follow tracked objects
+- Observation render was optimized